[#34] Refactor: 사용하지 않는 LargeButton과 CompleteLargeButton을 삭제하고 파일명을 LargeButton에서 LargeButtonStyle로 변경
